requirements.txt
streamlit==1.22.0
python-dotenv==1.0.0
pypdf==3.12.0
langchain==0.0.123
requests==2.31.0
faiss-cpu==1.7.3
torch==2.0.1
transformers==4.30.2
sentence-transformers==2.2.2
